Pull replyServer_ from O1 and O2 up into O0BaseAuth
This change allows to show a custom HTML page after a successful
authorization using OAuth1. It also eliminates some code duplication.
