+++ src/o0baseauth.h
@@ -12,6 +12,8 @@
 #include "o0abstractstore.h"
 #include "o0requestparameter.h"
 
+class O2ReplyServer;
+
 /// Base class of OAuth authenticators
 class O0_EXPORT O0BaseAuth : public QObject {
     Q_OBJECT
@@ -47,6 +49,12 @@ public:
     Q_PROPERTY(QString clientSecret READ clientSecret WRITE setClientSecret NOTIFY clientSecretChanged)
     QString clientSecret();
     void setClientSecret(const QString &value);
+
+    /// Page content on local host after successful oauth.
+    /// Provide it in case you do not want to close the browser, but display something
+    Q_PROPERTY(QByteArray replyContent READ replyContent WRITE setReplyContent)
+    QByteArray replyContent() const;
+    void setReplyContent(const QByteArray &value);
 
     /// TCP port number to use in local redirections.
     /// The OAuth "redirect_uri" will be set to "http://localhost:<localPort>/".
@@ -116,6 +124,7 @@ protected:
     quint16 localPort_;
     O0AbstractStore *store_;
     QVariantMap extraTokens_;
+    O2ReplyServer *replyServer_;
 };
 
 #endif // O0BASEAUTH
